requirements_jetson.txt
# VisionPilot XR - Requirements for NVIDIA Jetson Orin Nano
#
# Installation:
#   pip install -r requirements_jetson.txt
#
# Note: PyTorch must be installed separately with ARM64 wheels:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0

# GUI
PyQt5>=5.15.0
PyQt5-sip>=12.0

# Serial communication
pyserial>=3.5
pyserial-asyncio>=0.6

# Camera
pyrealsense2>=2.54.0

# System monitoring
psutil>=5.9.0

# Machine learning (already installed via pip separately)
# torch>=2.1.0  -> Requires: pip install torch --index-url https://download.pytorch.org/whl/cu121
# torchvision>=0.16.0

# Optional but recommended
jetson-stats>=4.2.0  # Monitor Jetson performance

# Development tools (optional)
# jupyter>=1.0.0
# ipython>=8.0.0

